refactor: integration with 3D viewer
