+++ frontend/src/components/explorer/mapViewer/ThreeDviewer.vue
@@ -1,24 +1,13 @@
 <template>
-  <div class="extended-section">
-    <div class="columns">
-      <template v-if="errorMessage">
-        <div class="column">
-          <div class="columns">
-            <div class="column notification is-danger is-half is-offset-one-quarter has-text-centered">
-              <p>{{ errorMessage }}</p>
-            </div>
-          </div>
-        </div>
-      </template>
-      <template v-else>
-        <div id="viewer" class="atlas-viewer"></div>
-      </template>
+  <div v-if="errorMessage" class="columns is-centered">
+    <div class="column notification is-danger is-half is-offset-one-quarter has-text-centered">
+      {{ errorMessage }}
     </div>
   </div>
+  <div v-else id="3dviewer"></div>
 </template>
 
 <script>
-
 import { mapState } from 'vuex';
 import { MetAtlasViewer } from '@metabolicatlas/mapviewer-3d';
 import { default as messages } from '@/helpers/messages';
@@ -26,8 +15,10 @@ import { default as messages } from '@/helpers/messages';
 export default {
   name: 'ThreeDViewer',
   props: {
-    componentType: String,
-    componentId: String,
+    componentId: {
+      type: String,
+      required: true,
+    },
     loading: Boolean,
   },
   data() {
@@ -44,38 +35,31 @@ export default {
     }),
   },
   watch: {
-    async componentType() {
-      this.resetNetwork();
-      await this.loadNetwork();
-    },
     async componentId() {
       this.resetNetwork();
       await this.loadNetwork();
     },
   },
+  async mounted() {
+    await this.loadNetwork();
+  },
   methods: {
     async loadNetwork() {
-      if (this.loading) {
-        return; // prevent duplicate loads
-      }
-
       const payload = {
         model: this.model.apiName,
         version: this.model.apiVersion,
-        type: this.componentType,
+        type: 'subsystem',
         id: this.componentId,
       };
-
       await this.$store.dispatch('maps/get3DMapNetwork', payload);
       this.renderNetwork();
-      // console.log('controller:', controller);
       // controller.filterBy({group: 'm'});
       // controller.filterBy({id: [1, 2, 3, 4]});
       // Subscribe to node selection events
       // document.getElementById('viewer').addEventListener('select', e => console.debug('selected', e.detail));
     },
     renderNetwork() {
-      this.controller = MetAtlasViewer('viewer');
+      this.controller = MetAtlasViewer('3dviewer');
       this.controller.setData(
         this.network,
         [{ group: 'e', sprite: '/sprite_round.png' },
@@ -84,7 +68,7 @@ export default {
         15);
     },
     resetNetwork() {
-      const viewer = document.getElementById('viewer');
+      const viewer = document.getElementById('3dviewer');
       viewer.innerHTML = '';
     },
   },
@@ -92,10 +76,8 @@ export default {
 </script>
 
 <style lang='scss'>
-.atlas-viewer {
-  width: 100%;
-  height: 100%;
-  margin: 0;
-  padding: 0;
+#3dviewer {
+  min-width: 100%;
+  min-height: 100%;
 }
 </style>
